backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.32
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
openai==1.40.1
alembic==1.13.2
bcrypt==4.0.1
httpx==0.27.2
